+++ source/components/Tooltip.js
@@ -7,7 +7,7 @@ import { createEmptyContext, withTheme } from './HOC/withTheme';
 import { composeTheme, addThemeId, didThemePropsChange } from '../utils/themes';
 
 // import constants
-import { IDENTIFIERS } from '../themes/API';
+import { IDENTIFIERS } from '.';
 import type { ThemeContextProp } from './HOC/withTheme';
 
 type Props = {
@@ -17,7 +17,7 @@ type Props = {
   isBounded?: boolean,
   isOpeningUpward: boolean,
   isTransparent: boolean,
-  skin: ComponentType<any>,
+  skin?: ComponentType<any>,
   theme: ?Object, // will take precedence over theme in context if passed
   themeOverrides: Object, // custom css/scss from user that adheres to component's theme API
   themeId: string,
@@ -60,7 +60,9 @@ class TooltipBase extends Component<Props, State> {
 
   render() {
     // destructuring props ensures only the "...rest" get passed down
-    const { skin: TooltipSkin, theme, themeOverrides, context, ...rest } = this.props;
+    const { skin, theme, themeOverrides, context, ...rest } = this.props;
+
+    const TooltipSkin = skin || context.skins[IDENTIFIERS.TOOLTIP];
 
     return <TooltipSkin theme={this.state.composedTheme} {...rest} />;
   }